end of project

+++ libft/incs/libft.h
@@ -6,7 +6,7 @@
 /*   By: mtrullar <mtrullar@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/20 13:32:30 by mtrullar          #+#    #+#             */
-/*   Updated: 2024/07/22 16:54:55 by mtrullar         ###   ########.fr       */
+/*   Updated: 2024/07/24 16:40:15 by mtrullar         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,7 +23,7 @@ typedef struct s_list
 	struct s_list	*next;
 }					t_list;
 
-int		ft_atoi(const char *nptr);
+long	ft_atoi(const char *nptr);
 
 void	ft_bzero(void *s, size_t n);
